Rename createuser script to avoid conflict with PostgreSQL client
